+++ Medical Inventory/Data/IRepository/ICategoryRepository.cs
@@ -5,7 +5,7 @@ namespace Medical_Inventory.Data.IRepository;
 public interface ICategoryRepository : IRepository<Category>
 {
     void Update(Category obj);
-    void Remove(int id);
+    void Remove(long id);
     Task<Category?> GetByName(string name);
 
 }
